Modify DeviceType trigger
Use platform-specific matching APIs and names
