+++ src/WindowsStateTriggers/WindowsStateTriggers.csproj
@@ -116,7 +116,7 @@
     <Compile Include="IsTrueStateTrigger.cs" />
     <Compile Include="OrientationStateTrigger.cs" />
     <Compile Include="IsTypePresentStateTrigger.cs" />
-    <Compile Include="DeviceTypeAdaptiveTrigger.cs" />
+    <Compile Include="DeviceFamilyAdaptiveTrigger.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <PropertyGroup Condition=" '$(VisualStudioVersion)' == '' or '$(VisualStudioVersion)' &lt; '14.0' ">
